requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
cssselect2==0.7.0
Django==5.0.6
django-admin-interface==0.28.6
django-appconf==1.0.6
django-autoslug==1.9.9
django-browser-reload==1.12.1
django-cities-light==3.10.1
django-ckeditor-5==0.2.13
django-colorfield==0.11.0
django-compressor==4.4
django-libsass==0.9
django-model-utils==4.5.1
django-notifications-hq==1.8.3
fonttools==4.53.0
html5lib==1.1
idna==3.7
jsonfield==3.1.0
lxml==5.2.2
numpy==1.26.4
oscrypto==1.3.0
pandas==2.2.2
pdfkit==1.0.0
pillow==10.3.0
progressbar2==4.4.2
psycopg2==2.9.9
pycparser==2.22
pydyf==0.10.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rcssmin==1.1.1
reportlab==4.0.9
requests==2.32.3
rjsmin==1.2.1
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
swapper==1.3.0
text-unidecode==1.3
tinycss2==1.3.0
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
Unidecode==1.3.8
uritools==4.0.3
urllib3==2.2.1
weasyprint==62.2
webencodings==0.5.1
xhtml2pdf==0.2.15
zopfli==0.2.3